requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.9.3
certifi==2022.9.24
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.0.4
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
geographiclib==2.0
geopy==2.3.0
googlemaps==4.7.2
googlesearch-python==1.1.0
greenlet==2.0.1
h11==0.14.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
numpy==1.23.5
outcome==1.2.0
pandas==1.5.1
protobuf==3.20.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
psycopg2==2.9.5
requests==2.25.1
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
Werkzeug==2.2.2
wsproto==1.2.0
